tf_cnn_model_load_testing/requirements.txt
absl-py==0.9.0
appdirs==1.4.3
astunparse==1.6.3
CacheControl==0.12.6
cachetools==4.1.0
certifi==2019.11.28
chardet==3.0.4
click==7.1.2
colorama==0.4.3
ConfigArgParse==1.2.3
contextlib2==0.6.0
distlib==0.3.0
distro==1.4.0
Flask==1.1.2
Flask-BasicAuth==0.2.0
gast==0.3.3
gevent==20.4.0
geventhttpclient-wheels==1.3.1.dev2
google-auth==1.14.3
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
greenlet==0.4.15
grpcio==1.29.0
gunicorn==20.0.4
h5py==2.10.0
html5lib==1.0.1
idna==2.8
ipaddr==2.2.0
itsdangerous==1.1.0
Jinja2==2.11.2
Keras-Preprocessing==1.1.2
lockfile==0.12.2
locust==1.0.1
locustio==0.14.6
Markdown==3.2.2
MarkupSafe==1.1.1
msgpack==0.6.2
numpy==1.18.4
oauthlib==3.1.0
opt-einsum==3.2.1
packaging==20.3
pep517==0.8.2
Pillow==7.1.2
progress==1.5
protobuf==3.12.0
psutil==5.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.6
pytoml==0.1.21
pyzmq==19.0.1
requests==2.22.0
requests-oauthlib==1.3.0
retrying==1.3.3
rsa==4.0
scipy==1.4.1
six==1.12.0
tensorboard==2.2.1
tensorboard-plugin-wit==1.6.0.post3
tensorflow==2.2.0
tensorflow-estimator==2.2.0
termcolor==1.1.0
urllib3==1.25.8
webencodings==0.5.1
Werkzeug==1.0.1
wrapt==1.12.1
